+++ src/main/java/com/refinedmods/refinedstorage/apiimpl/storage/tracker/ItemStorageTracker.java
@@ -2,7 +2,7 @@ package com.refinedmods.refinedstorage.apiimpl.storage.tracker;
 
 import com.refinedmods.refinedstorage.api.storage.tracker.IStorageTracker;
 import com.refinedmods.refinedstorage.api.storage.tracker.StorageTrackerEntry;
-import com.refinedmods.refinedstorage.apiimpl.API;
+import com.refinedmods.refinedstorage.util.ItemStackKey;
 import com.refinedmods.refinedstorage.util.StackUtils;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
@@ -17,7 +17,7 @@ public class ItemStorageTracker implements IStorageTracker<ItemStack> {
     private static final String NBT_NAME = "Name";
     private static final String NBT_TIME = "Time";
 
-    private final Map<Key, StorageTrackerEntry> changes = new HashMap<>();
+    private final Map<ItemStackKey, StorageTrackerEntry> changes = new HashMap<>();
     private final Runnable listener;
 
     public ItemStorageTracker(Runnable listener) {
@@ -26,14 +26,14 @@ public class ItemStorageTracker implements IStorageTracker<ItemStack> {
 
     @Override
     public void changed(PlayerEntity player, ItemStack stack) {
-        changes.put(new Key(stack), new StorageTrackerEntry(System.currentTimeMillis(), player.getName().getString()));
+        changes.put(new ItemStackKey(stack), new StorageTrackerEntry(System.currentTimeMillis(), player.getName().getString()));
 
         listener.run();
     }
 
     @Override
     public StorageTrackerEntry get(ItemStack stack) {
-        return changes.get(new Key(stack));
+        return changes.get(new ItemStackKey(stack));
     }
 
     public void readFromNbt(ListNBT list) {
@@ -43,7 +43,7 @@ public class ItemStorageTracker implements IStorageTracker<ItemStack> {
             ItemStack stack = StackUtils.deserializeStackFromNbt(tag.getCompound(NBT_STACK));
 
             if (!stack.isEmpty()) {
-                changes.put(new Key(stack), new StorageTrackerEntry(tag.getLong(NBT_TIME), tag.getString(NBT_NAME)));
+                changes.put(new ItemStackKey(stack), new StorageTrackerEntry(tag.getLong(NBT_TIME), tag.getString(NBT_NAME)));
             }
         }
     }
@@ -51,34 +51,16 @@ public class ItemStorageTracker implements IStorageTracker<ItemStack> {
     public ListNBT serializeNbt() {
         ListNBT list = new ListNBT();
 
-        for (Map.Entry<Key, StorageTrackerEntry> entry : changes.entrySet()) {
+        for (Map.Entry<ItemStackKey, StorageTrackerEntry> entry : changes.entrySet()) {
             CompoundNBT tag = new CompoundNBT();
 
             tag.putLong(NBT_TIME, entry.getValue().getTime());
             tag.putString(NBT_NAME, entry.getValue().getName());
-            tag.put(NBT_STACK, StackUtils.serializeStackToNbt(entry.getKey().stack));
+            tag.put(NBT_STACK, StackUtils.serializeStackToNbt(entry.getKey().getStack()));
 
             list.add(tag);
         }
 
         return list;
-    }
-
-    private static class Key {
-        private final ItemStack stack;
-
-        public Key(ItemStack stack) {
-            this.stack = stack;
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            return other instanceof Key && API.instance().getComparer().isEqualNoQuantity(stack, ((Key) other).stack);
-        }
-
-        @Override
-        public int hashCode() {
-            return API.instance().getItemStackHashCode(stack);
-        }
     }
 }
